% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{add_concentration}
\alias{add_concentration}
\title{Add concentrations}
\usage{
add_concentration(df_layout, concentrations = 10^(seq(-3, 1, 0.5)))
}
\arguments{
\item{df_layout}{data.table that should contains the cell line,
drug, concentration, and replicate columns along with the annotations that needs to be propagated}

\item{concentrations}{vector of numeric concentrations that will be added to df_layout}
}
\value{
data.table with concentrations
}
\description{
Add concentrations
}
\examples{

cell_lines <- create_synthetic_cell_lines()
add_concentration(cell_lines)

}
