% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{add_data_replicates}
\alias{add_data_replicates}
\title{Add data replicates}
\usage{
add_data_replicates(df_layout)
}
\arguments{
\item{df_layout}{data.frame that should contains the cell line,
drug, concentration, and replicate columns along with the annotations that needs to be propagated}
}
\value{
data.frame with replicates
}
\description{
Add data replicates
}
