% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{cell_lines}
\alias{cell_lines}
\title{Cell lines}
\value{
data.table
}
\description{
Cell lines
}
\examples{
path <- system.file("annotation_data", "cell_lines.csv", package = "gDRtestData")
data.table::fread(file = path)

}
\keyword{data}
\keyword{internal}
