% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{DepMap_Model}
\alias{DepMap_Model}
\title{DepMap Models (Cell Line Metadata)}
\format{
Table
\describe{
\item{ModelID}{Unique cell line identifier (ACH-XXXXX format)}
\item{CCLEName}{Cell line name from Cancer Cell Line Encyclopedia (CCLE)}
\item{CellLineName}{Common cell line name (e.g., "A375", "HCT-116")}
\item{TissueOrigin}{Indicates tissue model was derived from (Human, Mouse, Other)}
\item{DepmapModelType}{Abbreviated ID for model type. For cancer models, from Oncotree;
for other disease types, generated by DepMap}
\item{OncotreeLineage}{Primary cancer lineage classification (from Oncotree for cancer models,
generated by DepMap for others)}
\item{OncotreePrimaryDisease}{Primary disease of model (from Oncotree for cancer models,
generated by DepMap for others)}
\item{OncotreeSubtype}{Subtype classification of model (from Oncotree for cancer models,
generated by DepMap for others)}
\item{OncotreeCode}{Oncotree code for cancer models; left blank if
no corresponding code exists}
\item{PrimaryOrMetastasis}{Site of tumor origin (Primary/Metastatic/Recurrence/Other/Unknown)}
\item{Age}{Age at time of tissue sampling}
\item{AgeCategory}{Age category at time of sampling (Adult/Pediatric/Fetus/Unknown)}
\item{Sex}{Sex at time of sampling (Female/Male/Unknown)}
\item{PatientRace}{Patient/clinical indicated race (not derived)}
}
}
\source{
\url{https://depmap.org/portal/data_page/?tab=allData}
}
\description{
Cell line information from DepMap Public 24Q4.
}
\details{
Downloaded May 26, 2024 from DepMap Portal (version 24Q4).

Citation: DepMap, Broad (2024). DepMap 24Q4 Public. Figshare+.
https://doi.org/10.25452/figshare.plus.27993248.v1
}
\keyword{internal}
