% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{DepMap_OmicsExpressionProteinCodingGenesTPMLogp1}
\alias{DepMap_OmicsExpressionProteinCodingGenesTPMLogp1}
\title{Gene Expression (Log2 TPM)}
\format{
Matrix with ~1000 rows (cell lines) and ~19,000 columns (genes)
}
\source{
\url{https://depmap.org/portal/data_page/?tab=allData}
}
\description{
RNA-seq based gene expression for protein-coding genes.
}
\details{
\itemize{
\item Rows: Cell line identifiers
\item Columns: NCBI gene IDs (Entrez format)
\item Values: Log2(TPM + 1) transformed expression
\item Only protein-coding genes included
}
}
\keyword{internal}
