% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{drugs}
\alias{drugs}
\title{Drugs}
\value{
data.table
}
\description{
Drugs
}
\examples{
path <- system.file("annotation_data", "drugs.csv", package = "gDRtestData")
data.table::fread(file = path)

}
\keyword{data}
\keyword{internal}
