% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRtestData-package.R
\docType{package}
\name{gDRtestData-package}
\alias{gDRtestData}
\alias{gDRtestData-package}
\title{gDRtestData: gDRtestData - R data package with testing dose response data}
\value{
package help page
}
\description{
R package with internal dose-response test data. Package provides functions to generate input testing data that can be used as the input for gDR pipeline. It also contains qs2 files with MAE data processed by gDR.
}
\note{
To learn more about functions start with \code{help(package = "gDRtestData")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDRtestData}
  \item \url{https://gdrplatform.github.io/gDRtestData/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDRtestData/issues}
}

}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})

Authors:
\itemize{
  \item Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})
  \item Aleksander Chlebowski
  \item Bartosz Czech (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
  \item Marc Hafner (\href{https://orcid.org/0000-0003-1337-7598}{ORCID})
  \item Marcin Kamianowski
  \item Sergiu Mocanu
  \item Natalia Potocka
  \item Dariusz Scigocki
  \item Janina Smola (\href{https://orcid.org/0009-0007-4347-7748}{ORCID})
  \item Allison Voung
}

}
\keyword{internal}
