% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{generate_e_inf}
\alias{generate_e_inf}
\title{Calculate E inf metric}
\usage{
generate_e_inf(drugs, cell_lines, seed = 2)
}
\arguments{
\item{drugs}{data.frame with drugs}

\item{cell_lines}{data.frame with cell lines}

\item{seed}{Integer specifying the seed to set}
}
\value{
matrix with random E inf
}
\description{
Calculate E inf metric
}
