% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{generate_hill_coef}
\alias{generate_hill_coef}
\title{Generate hill coefficient}
\usage{
generate_hill_coef(drugs, cell_lines, seed = 2)
}
\arguments{
\item{drugs}{data.frame with drugs}

\item{cell_lines}{data.frame with cell lines}

\item{seed}{Integer specifying the seed to set}
}
\value{
matrix with random hill coefficient
}
\description{
Generate hill coefficient
}
