% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateCodilutionSmall}
\alias{generateCodilutionSmall}
\title{generateCodilutionSmall}
\usage{
generateCodilutionSmall(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateCodilutionSmall
}
\keyword{internal}
