% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboManyDrugs}
\alias{generateComboManyDrugs}
\title{generateComboManyDrugs}
\usage{
generateComboManyDrugs(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateComboManyDrugs
}
\keyword{internal}
