% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboMatrix}
\alias{generateComboMatrix}
\title{generateComboMatrix}
\usage{
generateComboMatrix(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateComboMatrix
}
\keyword{internal}
