% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboMatrixSmall}
\alias{generateComboMatrixSmall}
\title{generateComboMatrixSmall}
\usage{
generateComboMatrixSmall(
  cell_lines,
  drugs,
  e_inf,
  ec50,
  hill_coef,
  save = TRUE
)
}
\description{
generateComboMatrixSmall
}
\keyword{internal}
