% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateComboNoNoiseData2}
\alias{generateComboNoNoiseData2}
\title{generateComboNoNoiseData2}
\usage{
generateComboNoNoiseData2(
  cell_lines,
  drugs,
  e_inf,
  ec50,
  hill_coef,
  save = TRUE
)
}
\description{
generateComboNoNoiseData2
}
\keyword{internal}
