% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateLigandData}
\alias{generateLigandData}
\title{generateLigandData}
\usage{
generateLigandData(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateLigandData
}
\keyword{internal}
