% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateManyDrugsData}
\alias{generateManyDrugsData}
\title{generateManyDrugsData}
\usage{
generateManyDrugsData(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateManyDrugsData
}
\keyword{internal}
