% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateManyLinesData}
\alias{generateManyLinesData}
\title{generateManyLinesData}
\usage{
generateManyLinesData(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateManyLinesData
}
\keyword{internal}
