% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateMediumData}
\alias{generateMediumData}
\title{generateMediumData}
\usage{
generateMediumData(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateMediumData
}
\keyword{internal}
