% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateNoiseRawData}
\alias{generateNoiseRawData}
\title{generateNoiseRawData}
\usage{
generateNoiseRawData(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateNoiseRawData
}
\keyword{internal}
