% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateNoNoiseRawData}
\alias{generateNoNoiseRawData}
\title{generateNoNoiseRawData}
\usage{
generateNoNoiseRawData(cell_lines, drugs, e_inf, ec50, hill_coef, save = TRUE)
}
\description{
generateNoNoiseRawData
}
\keyword{internal}
