% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wrappers.R
\name{generateTripleComboMatrix}
\alias{generateTripleComboMatrix}
\title{generateTripleComboMatrix}
\usage{
generateTripleComboMatrix(
  cell_lines,
  drugs,
  e_inf,
  ec50,
  hill_coef,
  save = TRUE
)
}
\description{
generateTripleComboMatrix
}
\keyword{internal}
