% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_datasets.R
\name{get_test_dataset_paths}
\alias{get_test_dataset_paths}
\title{get_test_dataset_paths}
\usage{
get_test_dataset_paths()
}
\value{
named vector of absolute paths
}
\description{
Returns named vector of absolute paths to test datasets.
}
\author{
Kamil Foltyński \email{kamil.foltynski@contractors.roche.com}
}
