% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_datasets.R
\name{get_config_path}
\alias{get_config_path}
\title{get_config_path}
\usage{
get_config_path()
}
\value{
absolute path to config file
}
\description{
Return path to config in \code{yml} format containing list of test datasets
}
