% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepareCodilutionData}
\alias{prepareCodilutionData}
\title{prepareCodilutionData}
\usage{
prepareCodilutionData(df, df_layout)
}
\arguments{
\item{df}{data.table object with experiment data}

\item{df_layout}{data.table object with experiment design containg drugs and cell line data}
}
\value{
data.table with input data for testing
}
\description{
Create data.table with input codilution data containg noise for testing purpses
}
\examples{

df_layout <- prepareData(create_synthetic_cell_lines()[seq_len(2), ],
create_synthetic_drugs()[seq_len(4), ])
df <- cbind(create_synthetic_cell_lines()[1, , drop = FALSE],
df_layout[, "Concentration", drop = FALSE])
prepareCodilutionData(df, df_layout)

}
