% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepareComboMergedData}
\alias{prepareComboMergedData}
\title{prepareComboMergedData}
\usage{
prepareComboMergedData(
  cell_lines,
  drugs,
  drugsIdx1 = 2:4,
  drugsIdx2 = c(26, 26, 26),
  concentration = c(0, 0.2, 1),
  noise = 0.1,
  modifyDf2 = FALSE
)
}
\arguments{
\item{cell_lines}{data.table with cell line info}

\item{drugs}{data.table with drug info}

\item{drugsIdx1}{numeric vector of ids for primary drug}

\item{drugsIdx2}{numeric vector of ids for secondary drug}

\item{concentration}{numeric vector of doses}

\item{noise}{number indicating level of noise}

\item{modifyDf2}{Boolean indicating if the table should me modified to keep reverse
single agent data}
}
\value{
data.table with input data for testing
}
\description{
Create data.table with input commbination data containg noise for testing purpses
}
\examples{

prepareComboMergedData(create_synthetic_cell_lines(), create_synthetic_drugs())

}
