% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(cell_lines, drugs, conc = 10^(seq(-3, 1, 0.5)))
}
\arguments{
\item{cell_lines}{data.table with cell line info}

\item{drugs}{data.table with drug info}

\item{conc}{vector of doses}
}
\value{
data.table with input data for testing
}
\description{
Create data.table with input data for testing purposes
}
\examples{

prepareData(create_synthetic_cell_lines(), create_synthetic_drugs())

}
\keyword{helper_functions}
