% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepareMergedData}
\alias{prepareMergedData}
\title{prepareMergedData}
\usage{
prepareMergedData(cell_lines, drugs, noise = 0.1)
}
\arguments{
\item{cell_lines}{data.table with cell line info}

\item{drugs}{data.table with drug info}

\item{noise}{number indicating level of noise}
}
\value{
data.table with input data for testing
}
\description{
Create data.table with input data containing noise for testing purposes
}
\examples{

prepareMergedData(create_synthetic_cell_lines(), create_synthetic_drugs())

}
\keyword{helper_functions}
