% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-utils.R
\name{test_accuracy}
\alias{test_accuracy}
\title{Test accuracy of the fitted metrics based on the model}
\usage{
test_accuracy(se, e_inf, ec50, hill_coef)
}
\arguments{
\item{se}{SummarizedExperiment for testing}

\item{e_inf}{e_inf}

\item{ec50}{ec50}

\item{hill_coef}{hill_coef}
}
\value{
data.frame with quality metrics
}
\description{
Test accuracy of the fitted metrics based on the model
}
