% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{add_day0_data}
\alias{add_day0_data}
\title{Add data with day 0}
\usage{
add_day0_data(df_merged, noise_level = 0.05)
}
\arguments{
\item{df_merged}{data.frame with merged data}

\item{noise_level}{numeric scalar with the level of noise added to the data}
}
\description{
Add data with day 0
}
