% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{add_day0_data}
\alias{add_day0_data}
\title{Add data with day 0}
\usage{
add_day0_data(df_merged, noise_level = 0.05)
}
\arguments{
\item{df_merged}{data.table with merged data}

\item{noise_level}{numeric scalar with the level of noise added to the data}
}
\value{
data.table with day0 data
}
\description{
Add data with day 0
}
\examples{

cell_lines <- create_synthetic_cell_lines()
drugs <- create_synthetic_drugs()
data <- gDRtestData:::prepareData(cell_lines[seq_len(2), ], drugs[seq_len(4), ])
data$Duration <- 72
data$ReadoutValue <- 0
add_day0_data(data)

}
