% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{create_synthetic_cell_lines}
\alias{create_synthetic_cell_lines}
\title{Create data.table with synthetic cell lines}
\usage{
create_synthetic_cell_lines()
}
\value{
data.table with synthetic cell lines
}
\description{
Create data.table with synthetic cell lines
}
\examples{
create_synthetic_cell_lines()

}
\keyword{random_data}
