% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{create_synthetic_drugs}
\alias{create_synthetic_drugs}
\title{Create data.table with synthetic drugs}
\usage{
create_synthetic_drugs()
}
\value{
data.table with synthetic drugs
}
\description{
Create data.table with synthetic drugs
}
\examples{
create_synthetic_drugs()

}
\keyword{generate_data}
