% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{DepMap_OmicsExpressionProteinCodingGenesTPMLogp1}
\alias{DepMap_OmicsExpressionProteinCodingGenesTPMLogp1}
\title{Gene Expression (Log2 TPM)}
\format{
Matrix with cell lines as rows and genes as column
}
\description{
RNA-seq based gene expression for protein-coding genes.
}
\section{Loading}{

\code{data.table::fread(
system.file("depmap_data/OmicsExpressionProteinCodingGenesTPMLogp1.csv.gz",
package = "gDRtestData"))}
}

\section{Description}{

\itemize{
\item Rows: Cell line identifiers
\item Columns: NCBI gene IDs (Entrez format)
\item Values: Log2(TPM + 1) transformed expression
\item Only protein-coding genes included
}
}

\section{Source}{

\href{https://depmap.org/portal/data_page/?tab=allData}{DepMap Portal - Data}
}

\section{Details}{

Downloaded May 26, 2026 from DepMap Portal (version 24Q4).

Citation: 24Q4 DepMap Release, including CRISPR Screens, PRISM Drug Screens, Copy Number,
Mutation, Expression, and Fusions
DepMap, Broad (2024). DepMap 24Q4 Public. Figshare+. Dataset.
\url{https://doi.org/10.25452/figshare.plus.27993248.v1}
}

\keyword{real_data}
