% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{DepMap_OmicsSomaticMutationsMatrixHotspot}
\alias{DepMap_OmicsSomaticMutationsMatrixHotspot}
\title{Somatic Mutations (Hotspot)}
\format{
Matrix with cell lines as rows and genes as columns
}
\description{
Binary matrix of known hotspot mutations from whole exome sequencing.
}
\section{Loading}{

\code{data.table::fread(system.file("depmap_data/OmicsSomaticMutationsMatrixHotspot.csv.gz",
package = "gDRtestData"))}
}

\section{Description}{

\itemize{
\item Rows: Cell line identifiers
\item Columns: NCBI gene IDs
\item Values: Binary (0 = no mutation, 1 = hotspot mutation)
\item Hotspots defined by COSMIC and OncoKB
}
}

\section{Source}{

\href{https://depmap.org/portal/data_page/?tab=allData}{DepMap Portal - Data}
}

\section{Details}{

Downloaded May 26, 2026 from DepMap Portal (version 24Q4).

Citation: 24Q4 DepMap Release, including CRISPR Screens, PRISM Drug Screens, Copy Number,
Mutation, Expression, and Fusions
DepMap, Broad (2024). DepMap 24Q4 Public. Figshare+. Dataset.
\url{https://doi.org/10.25452/figshare.plus.27993248.v1}
}

\keyword{real_data}
