% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRtestData-package.R
\docType{package}
\name{gDRtestData-package}
\alias{gDRtestData}
\alias{gDRtestData-package}
\title{gDRtestData: gDRtestData - R data package with testing dose reponse data}
\value{
package help page
}
\description{
R package with internal dose-response test data. Package provides functions to generate input testing data that can be used as the input for gDR pipeline. It also contains RDS files with MAE data processed by gDR.
}
\note{
To learn more about functions start with \code{help(package = "gDRtestData")}
}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com}

Authors:
\itemize{
  \item Aleksander Chlebowski
  \item Bartosz Czech
  \item Marc Hafner
  \item Marcin Kamianowski
  \item Sergiu Mocanu
  \item Natalia Potocka
  \item Dariusz Scigocki
  \item Allison Voung
}

}
\keyword{internal}
