% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{generate_e_inf}
\alias{generate_e_inf}
\title{Calculate E inf metric}
\usage{
generate_e_inf(drugs, cell_lines)
}
\arguments{
\item{drugs}{data.table with drugs}

\item{cell_lines}{data.table with cell lines}
}
\value{
matrix with random E inf
}
\description{
Calculate E inf metric
}
\examples{
generate_e_inf(create_synthetic_drugs(), create_synthetic_cell_lines())

}
\keyword{generate_results}
