% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{generate_hill_coef}
\alias{generate_hill_coef}
\title{Generate hill coefficient}
\usage{
generate_hill_coef(drugs, cell_lines)
}
\arguments{
\item{drugs}{data.table with drugs}

\item{cell_lines}{data.table with cell lines}
}
\value{
matrix with random hill coefficient
}
\description{
Generate hill coefficient
}
\examples{
generate_hill_coef(create_synthetic_drugs(), create_synthetic_cell_lines()) 

}
\keyword{random_data}
