% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_datasets.R
\name{get_test_dataset_paths}
\alias{get_test_dataset_paths}
\title{get_test_dataset_paths}
\usage{
get_test_dataset_paths(datasets_dir = NULL, config_path = NULL)
}
\arguments{
\item{datasets_dir}{path to directory with datasets (default \code{NULL}).
If \code{NULL}, then \code{inst/testdata} directory from \code{gDRtestData} will be used.}

\item{config_path}{path to config i \code{yml} format (default \code{NULL}).
If \code{NULL}, then config from \link{get_config_path} will be used.}
}
\value{
named vector of absolute paths
}
\description{
Returns named vector of absolute paths to test datasets.
}
\author{
Kamil Foltyński \email{kamil.foltynski@contractors.roche.com}
}
