% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prepareCodilutionData}
\alias{prepareCodilutionData}
\title{prepareCodilutionData}
\usage{
prepareCodilutionData(
  cell_lines,
  drugs,
  drugsIdx2 = 1,
  conc = 10^(seq(-3, 1, 0.5)),
  noise = 0.1
)
}
\arguments{
\item{cell_lines}{data.table with cell line info}

\item{drugs}{data.table with drug info}

\item{drugsIdx2}{numeric vector of ids for secondary drug (in \code{drugs} data.table)}

\item{conc}{vector of doses}

\item{noise}{number indicating level of noise}
}
\value{
data.table with input data for testing
}
\description{
Create data.table with input co-dilution data containing noise for testing purposes
}
\examples{

prepareCodilutionData(create_synthetic_cell_lines()[seq_len(2), ],
                      create_synthetic_drugs()[seq_len(4), ])

}
\keyword{helper_functions}
